LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/importScripts-expected.txt


FAIL Base URL in workers: importScripts assert_equals: expected "gamma" but got "beta"
